pyproject.toml
[tool.poetry]
name = "django-server"
version = "0.1.0"
description = ""
authors = ["daniel <ehdgur8584@gmail.com>"]
readme = "README.md"
packages = [{include = "django_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
shortuuid = "^1.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
